Create 小青的數學作業.C
So far, I've finished part of the program that divides mathematical formulas with parentheses.

When I finished it, I found that it has exceeded the limit, but I will continue with the current design, because it will be more perfect and I don't want to write the program just to cope with the teacher.